requirements.txt
pykrx>=1.0.45
openai>=1.12.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ta>=0.11.0
python-dotenv>=1.0.0
